Move "bar-getting from google" to GetBarListGoogle.
+ Remove unecessary imports
